+++ packages/ricos-content/web/src/RicosContentAPI/extract-by-type.spec.ts
@@ -48,6 +48,7 @@ describe('Extract by Types', () => {
   it('should run getMedia on all plugins', () => {
     expect(getMedia(content)).toStrictEqual([
       {
+        duration: undefined,
         src: {
           url: undefined,
           custom: undefined,
@@ -58,6 +59,7 @@ describe('Extract by Types', () => {
         height: 5577,
       },
       {
+        duration: undefined,
         src: {
           url: undefined,
           custom: undefined,
@@ -68,6 +70,7 @@ describe('Extract by Types', () => {
         height: 3727,
       },
       {
+        duration: 666,
         src: {
           url: undefined,
           custom: undefined,
@@ -78,6 +81,7 @@ describe('Extract by Types', () => {
         height: undefined,
       },
       {
+        duration: 666,
         src: {
           url: 'https://www.youtube.com/watch?v=iecNmOXDOHM',
           custom: undefined,
@@ -88,6 +92,7 @@ describe('Extract by Types', () => {
         height: undefined,
       },
       {
+        duration: 666,
         src: {
           url: undefined,
           custom: undefined,
@@ -98,6 +103,7 @@ describe('Extract by Types', () => {
         height: undefined,
       },
       {
+        duration: undefined,
         src: {
           url: '8bb438_2ecb6a1a9d5b4e148bd83a1ea3920860.jpg',
           custom: undefined,
@@ -108,6 +114,7 @@ describe('Extract by Types', () => {
         height: 3737,
       },
       {
+        duration: undefined,
         src: {
           url: '8bb438_131e9bbd900c42a29ed2db01d18256e5.jpg',
           custom: undefined,
@@ -118,6 +125,7 @@ describe('Extract by Types', () => {
         height: 1081,
       },
       {
+        duration: undefined,
         src: {
           url: '8bb438_a132b18dea524d48a68a37f0075fcc1e.jpg',
           custom: undefined,